Switch to a more muted color for the visual hints
Based on feedback in
https://twitter.com/Vlachbild/status/1277678971157610496.  I realized
I was using red just because that's what I used in todoist-shortcuts.
I'm used to the color and it fits better there.  Roam Navigator is
intended to be visible more frequently by default, and so it makes
sense to have the visual hints blend in more